agents/pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [{ include = "agentic" }]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.13"
Flask = "^3.0.3"
openai = { version = "^1.35.13", allow-prereleases = true }
python-dotenv = "^1.0.1"
flask-socketio = "^5.3.6"
apscheduler = "^3.10.4"
flask-login = "^0.6.3"
Werkzeug = ">=3.0.0"
Flask-SQLAlchemy = "^3.0.3"
flask-session = "^0.8.0"
flask-jwt-extended = "^4.6.0"
flask-cors = "^4.0.1"
flask-limiter = "^3.7.0"
flask-talisman = "^1.1.0"
redis = "^5.0.7"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"